src/test/java/com/blogspot/ostas/leetcode/all/easy/squares_of_a_sorted_array/SolutionTest.java
package com.blogspot.ostas.leetcode.all.easy.squares_of_a_sorted_array;

import org.junit.jupiter.api.Test;

import static org.assertj.core.api.Assertions.assertThat;

class SolutionTest {

    /*
      Example 1:
    Input: nums = [-4,-1,0,3,10]
    Output: [0,1,9,16,100]
    Explanation: After squaring, the array becomes [16,1,0,9,100].
    After sorting, it becomes [0,1,9,16,100].
    Example 2:
    Input: nums = [-7,-3,2,3,11]
    Output: [4,9,9,49,121]
      Constraints:
    1 <= nums.length <= 104
    -104 <= nums[i] <= 104
    nums is sorted in non-decreasing order.
      Follow up: Squaring each element and sorting the new array is very trivial, could you find an O(n) solution using a different approach?
    */
    @Test
    void example_0() {
        var solution = new Solution();
        int[] nums = new int[]{-4, -1, 0, 3, 10};
        int[] expected = new int[]{0, 1, 9, 16, 100};
        var result = solution.sortedSquares(nums);
        assertThat(result).isEqualTo(expected);
    }

    @Test
    void example_1() {
        var solution = new Solution();
        int[] nums = new int[]{-7, -3, 2, 3, 11};
        int[] expected = new int[]{4, 9, 9, 49, 121};
        var result = solution.sortedSquares(nums);
        assertThat(result).isEqualTo(expected);
    }

}
